fix: open course promo codes application (#873)
Relates to TTHP-823

Move the whole logic of querying for available promo codes to backend (for obvious reasons, normal users shouldn't just be able to query for all promo codes in the system). The mechanic assumes that if you know the code then you can query for the details of it. Backend also applies additional validation like if given promo code has been already used.

Co-authored-by: Danijel Maksimovic <danijel.maksimovic@nearform.com>